v1 of code
Notes: 

|||| Gamepad #1: |||| ABXY have no use, they do not do anything since odometry is not configured. Its always 100% power unless left trigger is down, if left trigger is down then it is 40%. To set all power of motors to 0 (in an emergency for example) click left stick button and right stick button at the same time and all motors will go 0 power.

 ||||  Gamepad #2:  |||| 1. Right joystick should pull arm up and down at the same speed as last year, however it may be inversed values for stick up + down, be careful when testing since idk if the min and max of the servo position works, and it could break the servo, but it should work. 2. Left joystick sets the servo for WG to 0 and 0.5 , put that is a servo position that depends on how it was set up. Be careful when testing this since it could bump into something. Text me what values I should change it to. 3. If A is clicked then the mag servo will go at full speed, once A is released it will go to 0 power. 4. The controller map info was very hard to understand for this. The motor for AA is set to 0.35 power when b is clicked, and set to -0.35 when x is clicked. The SH0 and SH1 were very hard to understand what they were for, so I made them be controlled with 2 buttons. X sets them to 0, and Y (only function for y) is set to 0.75 power. Right trigger and left trigger should work fine. 
TLDR; this is the best I could do without seeing the robot and with the stuff that was written with controller mapping. text me what to change servo positions to and stuff like that so the robot can run decently by Wednesday. Also there could be errors so if it does not compile then send me a screenshot. There is one thing that could make everything mess up, but i don't know until it is tested (also lots of values will be inversed from what you said in controller mapping, so just tell m if you want to change any directions)